feat: add hooks features
Add hooks features for app install.

+++ libs/linglong/src/linglong/package_manager/package_manager.cpp
@@ -20,6 +20,7 @@
 #include "linglong/utils/configure.h"
 #include "linglong/utils/error/error.h"
 #include "linglong/utils/finally/finally.h"
+#include "linglong/utils/hooks.h"
 #include "linglong/utils/packageinfo_handler.h"
 #include "linglong/utils/serialize/json.h"
 #include "linglong/utils/transaction.h"
@@ -735,6 +736,12 @@ QVariantMap PackageManager::installFromLayer(const QDBusUnixFileDescriptor &fd,
                           << "after install" << packageRef.toString() << ":"
                           << ret.error().message();
           }
+
+          ret = executePostInstallHooks(*newRef);
+          if (!ret) {
+              taskRef.reportError(std::move(ret).error());
+              return;
+          }
       };
 
     auto refSpec =
@@ -761,6 +768,18 @@ QVariantMap PackageManager::installFromLayer(const QDBusUnixFileDescriptor &fd,
 QVariantMap PackageManager::installFromUAB(const QDBusUnixFileDescriptor &fd,
                                            const api::types::v1::CommonOptions &options) noexcept
 {
+    std::unique_ptr<utils::InstallHookManager> installHookManager =
+      std::make_unique<utils::InstallHookManager>();
+    auto ret = installHookManager->parseInstallHooks();
+    if (!ret) {
+        return toDBusReply(ret);
+    }
+
+    ret = installHookManager->executeInstallHooks(fd.fileDescriptor());
+    if (!ret) {
+        return toDBusReply(-1, "uab package signature verification failed.");
+    }
+
     auto uabRet = package::UABFile::loadFromFile(
       QString("/proc/%1/fd/%2").arg(getpid()).arg(fd.fileDescriptor()));
     if (!uabRet) {
@@ -1064,6 +1083,12 @@ QVariantMap PackageManager::installFromUAB(const QDBusUnixFileDescriptor &fd,
                                     "Failed to generate some cache.\n" + result.error().message());
                 return;
             }
+        }
+
+        auto ret = executePostInstallHooks(newAppRef);
+        if (!ret) {
+            taskRef.reportError(std::move(ret).error());
+            return;
         }
 
         transaction.commit();
@@ -1381,6 +1406,13 @@ void PackageManager::Install(PackageTask &taskContext,
         }
     }
 
+    auto ret = executePostInstallHooks(newRef);
+    if (!ret) {
+        taskContext.updateState(linglong::api::types::v1::State::Failed,
+                                "Failed to execute postInstall hooks.\n" + ret.error().message());
+        return;
+    }
+
     transaction.commit();
     taskContext.updateState(linglong::api::types::v1::State::Succeed,
                             "Install " + newRef.toString() + " success");
@@ -1648,6 +1680,11 @@ void PackageManager::Uninstall(PackageTask &taskContext,
     UninstallRef(taskContext, ref, removedModules);
     if (isTaskDone(taskContext.subState())) {
         return;
+    }
+
+    auto ret = executePostUninstallHooks(ref);
+    if (!ret) {
+        qWarning() << "failed to execute postUninstall hooks" << ret.error();
     }
 
     transaction.commit();
@@ -1923,8 +1960,21 @@ void PackageManager::pullDependency(PackageTask &taskContext,
                 return;
             }
 
+            auto ret = executePostInstallHooks(*runtime);
+            if (!ret) {
+                taskContext.updateState(linglong::api::types::v1::State::Failed,
+                                        LINGLONG_ERRV(ret).message());
+                return;
+            }
+
             transaction.addRollBack([this, runtimeRef = *runtime, module]() noexcept {
                 auto result = this->repo.remove(runtimeRef, module);
+                if (!result) {
+                    qCritical() << result.error();
+                    Q_ASSERT(false);
+                }
+
+                result = executePostUninstallHooks(runtimeRef);
                 if (!result) {
                     qCritical() << result.error();
                     Q_ASSERT(false);
@@ -1964,6 +2014,27 @@ void PackageManager::pullDependency(PackageTask &taskContext,
         if (isTaskDone(taskContext.subState())) {
             return;
         }
+
+        auto ret = executePostInstallHooks(*base);
+        if (!ret) {
+            taskContext.updateState(linglong::api::types::v1::State::Failed,
+                                    LINGLONG_ERRV(ret).message());
+            return;
+        }
+
+        transaction.addRollBack([this, baseRef = *base, module]() noexcept {
+            auto result = this->repo.remove(baseRef, module);
+            if (!result) {
+                qCritical() << result.error();
+                Q_ASSERT(false);
+            }
+
+            result = executePostUninstallHooks(baseRef);
+            if (!result) {
+                qCritical() << result.error();
+                Q_ASSERT(false);
+            }
+        });
     }
 
     transaction.commit();
@@ -2422,6 +2493,53 @@ auto PackageManager::GenerateCache(const QString &reference) noexcept -> QVarian
       .message = "",
     });
     return result;
+}
+
+utils::error::Result<void>
+PackageManager::executePostInstallHooks(const package::Reference &ref) noexcept
+{
+    LINGLONG_TRACE("execute post install hooks for: " + ref.toString());
+
+    std::unique_ptr<utils::InstallHookManager> installHookManager =
+      std::make_unique<utils::InstallHookManager>();
+    auto ret = installHookManager->parseInstallHooks();
+    if (!ret) {
+        return LINGLONG_ERR(ret);
+    }
+
+    auto layerDir = this->repo.getLayerDir(ref);
+    if (!layerDir) {
+        return LINGLONG_ERR(layerDir);
+    }
+
+    auto appPath = layerDir->absolutePath();
+
+    ret = installHookManager->executePostInstallHooks(ref.id.toStdString(), appPath.toStdString());
+    if (!ret) {
+        return LINGLONG_ERR(ret);
+    }
+
+    return LINGLONG_OK;
+}
+
+utils::error::Result<void>
+PackageManager::executePostUninstallHooks(const package::Reference &ref) noexcept
+{
+    LINGLONG_TRACE("execute post uninstall hooks for: " + ref.toString());
+
+    std::unique_ptr<utils::InstallHookManager> installHookManager =
+      std::make_unique<utils::InstallHookManager>();
+    auto ret = installHookManager->parseInstallHooks();
+    if (!ret) {
+        return LINGLONG_ERR(ret);
+    }
+
+    ret = installHookManager->executePostUninstallHooks(ref.id.toStdString());
+    if (!ret) {
+        return LINGLONG_ERR(ret);
+    }
+
+    return LINGLONG_OK;
 }
 
 } // namespace linglong::service
